+++ src/SynExtend.h
@@ -33,8 +33,14 @@ SEXP pseudoRandomSample(SEXP N);
 SEXP randomProjection(SEXP VEC, SEXP NONZERO, SEXP N, SEXP OUTDIM);
 SEXP seededPseudoRandomSample(SEXP N, SEXP SEED);
 
+/**** MoranI.c ****/
+SEXP MoransI(SEXP VALS, SEXP DIST, SEXP DIM);
 
 /**** CShuffle.c ****/
 void shuffleRInt(int *v, int *l);
 
 void shuffleRRepl(int *v, int *l);
+
+
+/**** NucleotideCounts.c ****/
+SEXP StringToNVDT(SEXP DNASTRING, SEXP REMOVEGAPS, SEXP EXTENDED, SEXP USEDNA);